% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProj.R
\name{bootstrap_conf_intervals}
\alias{bootstrap_conf_intervals}
\title{Bootstrap Confidence Intervals}
\usage{
bootstrap_conf_intervals(X, y, alpha = 0.05, n_bootstraps = 20)
}
\arguments{
\item{alpha}{\code{scaler} statistical significance value (default is 0.05)}

\item{n_bootstraps}{code{character} defines the number of times the bootstrap will re-sample the data (default is 20)}

\item{resp}{y \code{vector} of length n.}

\item{pred}{X \code{matrix} n number rows beta plus 1 number of columns with first column being entirely ones.}
}
\value{
A \code{vector} that contains the confidence intervals of the bootstrap
}
\description{
This function resamples the data and lets the user choose the significance level α to obtain for the 1−α confidence intervals for β, and the number of bootstraps which by default is 20.
}
\examples{
set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X\%*\%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)\%*\%X)\%*\%t(X)\%*\%y
bootstrap_conf_intervals(X, y)
}
\author{
Elena Gagliano

Helen Wu

Max Van Horn
}
