% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProj.R
\name{confmat}
\alias{confmat}
\title{Confusion Matrix}
\usage{
confmat(y_actual, y_pred, cutoff = 0.5)
}
\arguments{
\item{y_actual}{\code{vector} containing the actual values of y.}

\item{y_pred}{\code{vector} containing predicted values of y.}

\item{cutoff}{\code{scaler} creates a cutoff to convert values based on the threshold (default is 0.5)}
}
\value{
A \list{numeric} giving measured prevalence, accuracy, and sensitivity of the function
}
\description{
This function creates the confusion matrix that measures the accuracy of the function
}
\examples{
set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X\%*\%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)\%*\%X)\%*\%t(X)\%*\%y
y_pred <- log_reg(X,y)$y_pred
confmat(y, y_pred)
}
\author{
Elena Gagliano

Helen Wu

Max Van Horn
}
