% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProj.R
\name{log_reg}
\alias{log_reg}
\title{Logistic Regression Function}
\usage{
log_reg(X, y)
}
\arguments{
\item{pred}{X \code{matrix} n number rows beta plus 1 number of columns with first column being entirely ones.}

\item{resp}{y \code{vector} of length n.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{betas}{estimated beta coefficients}
     \item{y_pred}{predicted y-values using estimated coefficients}
     \item{y_actual}{original vector of y-values}
     \item{X}{original matrix of predictors}
}
}
\description{
Logistic Regression Function
}
\examples{
set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X\%*\%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)\%*\%X)\%*\%t(X)\%*\%y
log_reg(X, y)
}
\author{
Elena Gagliano

Helen Wu

Max Van Horn
}
