% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProj.R
\name{plot.my_b}
\alias{plot.my_b}
\title{Logistic Regression Plots}
\usage{
\method{plot}{my_b}(obj)
}
\arguments{
\item{output}{log_reg function}
}
\value{
A graph{numeric} plotting predicted y-values against actual y-values and a logistic regression curve fitted using transformed x values and beta coefficient estimates
}
\description{
This function uses logistic regression computations to plot a fitted logistic curve in the data
}
\examples{
set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X\%*\%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)\%*\%X)\%*\%t(X)\%*\%y
test <- log_reg(X, y)
plot(test)
}
\author{
Elena Gagliano

Helen Wu

Max Van Horn
}
