% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProj.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Confusion Matrix Plots}
\usage{
plot_metrics(X, y, cutoff_values = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{cutoff_values}{\code{vector} a sequence of values from 0.1 to 9.9 separated by 0.1}

\item{resp}{y \code{vector} of length n.}

\item{pred}{X \code{matrix} n number rows beta plus 1 number of columns with first column being entirely ones.}
}
\value{
A \plot{numeric} with values for the prevalence, accuracy, sensitivity, and specificity
}
\description{
This function plots a grid that includes all of the metrics measured by the confusion matrix
}
\examples{
set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X\%*\%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)\%*\%X)\%*\%t(X)\%*\%y
plot_metrics(X,y)
}
\author{
Elena Gagliano

Helen Wu

Max Van Horn
}
